Stop creating MANIFEST.jzon.bz2
After https://github.com/web-platform-tests/wpt/pull/33493 we can rely
on MANIFEST.json.xz being available, and it will be both smaller and
faster to decompress than MANIFEST.json.bz2.

There's no reason to use GitHub's (free) storage for this.

It should be possible to drop MANIFEST.json.gz as well, but keep it
around so that there's one format that's available for every merge_pr_*
release, in case someone depends on that.